chore: drop code which was only used by <=2023.1
