+++ plugin-core/src/main/resources/META-INF/appmap-core.xml
@@ -68,10 +68,6 @@
         <fileBasedIndex implementation="appland.index.ClassMapTypeIndex"/>
 
         <roots.watchedRootsProvider implementation="appland.index.AppMapWatchedRootsProvider"/>
-        <!-- For 2023.1 and earlier -->
-        <!--suppress PluginXmlValidity -->
-        <indexableFilesContributor implementation="appland.index.AppMapIndexableFilesContributor"/>
-        <!-- For 2023.2 and later -->
         <indexedRootsProvider implementation="appland.index.AppMapIndexedRootsSetContributor"/>
 
         <toolWindow id="applandToolWindow" anchor="right" secondary="false"
